requirements.txt
amqp==5.0.6
arabic-reshaper==2.1.3
asgiref==3.4.1
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
celery==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.0.9
distlib==0.3.2
Django==3.2.7
django-celery-beat==2.0.0
django-ckeditor==6.2.0
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-email-users==1.1.0
django-filter==2.4.0
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-shopping-cart==0.1
django-timezone-field==4.2.1
django-toast-messages==0.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
env==0.1.0
filelock==3.0.12
future==0.18.2
html5lib==1.1
idna==3.3
image==1.5.33
jsonfield==3.1.0
kombu==5.1.0
kwargs==1.0.1
Pillow==8.3.2
platformdirs==2.3.0
prompt-toolkit==3.0.20
pycodestyle==2.7.0
PyPDF2==1.26.0
pytesseract==0.3.8
python-bidi==0.4.2
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
razorpay==1.2.0
reportlab==3.6.1
requests==2.26.0
self==2020.12.3
signals==0.0.2
six==1.16.0
sqlparse==0.4.2
swapper==1.1.2.post1
toml==0.10.2
urllib3==1.26.7
vine==5.0.0
virtualenv==20.7.2
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.5
